+++ CruelerThanDAT/inc/BinaryHandler.h
