+++ CruelerThanDAT/inc/CRC32.h
