+++ CruelerThanDAT/inc/CTDModel.h
