+++ CruelerThanDAT/inc/CommonTypes.h
