+++ CruelerThanDAT/inc/FileNodes.h
@@ -4,10 +4,10 @@
 #define NOMINMAX
 #include <string>
 #include <vector>
-#include "../globals.h"
+#include "globals.h"
 #include <imgui.h>
 #include "BinaryHandler.h"
-#include "../Assets/CodIcons.h"
+#include "CodIcons.h"
 #include <windows.h>
 #include <fstream> 
 #include "Log.h"
@@ -15,14 +15,14 @@
 #include <cmath>
 #include <iostream>
 #include <functional>
-#include "Utility/CRC32.h"
-#include "Utility/HashDataContainer.h"
+#include "CRC32.h"
+#include "HashDataContainer.h"
 #include <sstream>
 #include <format>
-#include "Utility/ImGuiExtended.h"
-#include "Utility/WMB.h"
-#include "Utility/CTDModel.h"
-#include "Utility/UID.h"
+#include "ImGuiExtended.h"
+#include "WMB.h"
+#include "CTDModel.h"
+#include "UID.h"
 #include <d3dx9.h>
 #include <chrono>
 #include "fbxsdk.h"
