+++ CruelerThanDAT/inc/HashDataContainer.h
