+++ CruelerThanDAT/inc/globals.h
