+++ CruelerThanDAT/inc/resource.h
