+++ CruelerThanDAT/src/CRC32.cpp
