+++ CruelerThanDAT/src/FileNodes.cpp
